user route

+++ routes/api/users.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const User = require('../../models/user');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
+const auth = require('../../middleware/auth');
 const { check, validationResult } = require('express-validator');
 const config = require('config');
 const bigchaindriver = require('bigchaindb-driver');
@@ -71,7 +72,19 @@ router.post('/login',
   [
     check('email', 'Please include a valid Email').isEmail(),
     check('password', 'Password is required').exists()],
-  async (req, res) => {
+    check('amkaUser')
+        .isLength({ min: 11, max: 11 })
+        .withMessage('Amka must be 11 numbers')
+        .matches(/^[0-9]+$/)
+        .withMessage('Is not an Amka type')
+        .notEmpty()
+        .withMessage('Amka of User is required'),
+    check('birthday', 'Birthday of User is required')
+    .matches(/^(0[1-9]|[12][0-9]|3[01])\/(0[1-9]|1[0-2])\/([12][0-9]{3})$/)
+    .withMessage('Is not a date type dd/mm/yyyy')
+    .notEmpty()
+    .withMessage('Date of birth required')
+    , async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(400).json({ errors: errors.array() });
@@ -126,7 +139,11 @@ router.put('/:userId', auth,
         .withMessage('Is not an Amka type')
         .notEmpty()
         .withMessage('Amka of User is required'),
-    check('birthday', 'Birthday of User is required').notEmpty(), async (req,res)  =>  {
+    check('birthday', 'Birthday of User is required')
+    .matches(/^(0[1-9]|[12][0-9]|3[01])\/(0[1-9]|1[0-2])\/([12][0-9]{3})$/)
+    .withMessage('Is not a date type dd/mm/yyyy')
+    .notEmpty()
+    .withMessage('Date of birth required'), async (req,res)  =>  {
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
           return res.status(400).json({ errors: errors.array() });
@@ -144,13 +161,15 @@ router.put('/:userId', auth,
 })
 
 router.delete('/:userId', auth, async (req, res) => {
+   if(req.user.id = req.params.userId){
     User.findByIdAndRemove(req.params.userId)
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+      .then((resp) => {
+          res.statusCode = 200;
+          res.setHeader('Content-type', 'application/json');
+          res.json(resp);
+      }, (err) => next(err))
+      .catch((err) => next(err));
+  }
 });
 
 module.exports = router;
